requirements.txt
python-telegram-bot==22.1
pandas==2.1.4
yadisk==3.3.0
python-dotenv==1.0.0